requirements.txt
django-ipware==1.1.6
djangorestframework==3.6.4
